// src/components/input/types.ts
import { UseFormRegister, RegisterOptions, FieldError, DeepMap } from 'react-hook-form';

// file component
type InputProps = {
  label: {
    name: string,
    label: string,
  };
  register: UseFormRegister<any>;
  placeholder?: string;
  showError?: boolean;
  rules?: RegisterOptions;
  errors?: DeepMap<any, FieldError>
  type?: "text" | "password" | "date" | "time" | "number" | "datetime-local" | "search";
}
// file styles
type stylesState = {
  formControl: string;
  label: string;
  iconSearch?: string;
  input: string;
  error: string;
};
type stylesProps = {
  error: boolean,
  type: "text" | "password" | "date" | "time" | "number" | "datetime-local" | "search";
}

export type { InputProps, stylesState, stylesProps };